Move active status model to reside with user model
